att_requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0
pdfplumber==0.11.3
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
google-ai-generativelanguage==0.1.0
google-generativeai==0.1.0rc1
google-auth==2.33.0
google-api-core==2.19.1
google-cloud-storage==2.18.2
tzdata==2024.1
